Fix hard-coded paths
Even though hard-coded paths aren’t a good solution, this fixed the paths
to Fast Downward for the time being. A much better solution would be a
config.example.py that can be customized by the users to a config.py and
that is read by the planner. However, I don’t have the resources to help
with a real fix right now :).