Remove infinite timeout in conditional wait (#6)
The timeout argument to Condition.wait() is optional. If omitted, the
method will block without a time limit, which very likely is what was
intended by using an infinite timeout in the first place.

This fixes issues with Python 3, where a timeout of infinity makes the
method block forever. This behavior is different from Python 2 but seems
not to be documented in the official Python documentation.