Delete config.py (#5)
config.py is a system-specific file. Users create this file by copying
the config.example.py and adapting it to their environment. For this
reason, config.py itself should not be part of this repository.
Otherwise, users will get merge conflicts every time changes to
config.py are made upstream, which really is undesirable.

For this reason, this removes the config.py. It’s already excluded from
being tracked by Git with .gitignore, which should prevent it from being
added back at a later point in time.