Fix Condition.wait() for Python 2 and Python 3 (#7)
Condition.wait() internally acquires a lock. For some reason, this
method is not interruptible in Python 2 if Condition.wait() is called
without a timeout.

The implementation of threading in Python 3 is completely different and
makes Condition.wait() interruptible. Also, an infinite timeout leads to
overflow issues.

For more information, see the Python bugtracker discussion [1].

Because there is no single solution to make the condition variable work
in both Python 2 and 3, this adds a switch to the timeout option
depending on the Python version.

[1] https://bugs.python.org/issue8844